refactor: split large network/seed.py into submodules (partial #180)
Split the 3,572-line seed.py into focused submodules:

- seed/__init__.py - Re-exports public API for backward compatibility
- seed/config.py - SeedConfig, regional routing (COUNTRY_TO_CONTINENT, etc.)
- seed/health.py - HealthStatus, HealthState, HealthMonitor
- seed/sybil.py - SybilResistance, ReputationManager, CorrelationDetector, RateLimiter
- seed/registry.py - RouterRecord dataclass
- seed/revocation.py - SeedRevocationManager, SeedRevocationRecord
- seed/gossip.py - SeedPeerManager for seed-to-seed gossip
- seed/core.py - SeedNode class, create_seed_node, run_seed_node

Benefits:
- Each module has a single responsibility
- Easier to navigate and maintain
- No circular import issues
- All existing imports continue to work unchanged
- All 263 seed-related tests pass

Note: node.py was already decomposed in Issue #128, so no changes needed there.
