+++ tests/network/test_seed_gossip.py
@@ -575,7 +575,7 @@ class TestGossipRound:
             "routers": [peer_router],
         })
         
-        with patch("valence.network.seed.aiohttp.ClientSession") as mock_session:
+        with patch("valence.network.seed.gossip.aiohttp.ClientSession") as mock_session:
             mock_ctx = MagicMock()
             mock_ctx.__aenter__ = AsyncMock(return_value=mock_response)
             mock_ctx.__aexit__ = AsyncMock()
